+++ addOns/jruby/src/main/zapHomeFiles/scripts/templates/passive/Passive default template.rb
@@ -5,15 +5,15 @@
 
 require 'java'
 java_package 'org.zaproxy.zap.extension.pscan'
-java_import 'org.zaproxy.zap.extension.pscan.PassiveScript'
+java_import 'org.zaproxy.zap.extension.scripts.scanrules.PassiveScript'
 java_import 'org.zaproxy.zap.extension.pscan.PluginPassiveScanner'
-java_import 'org.zaproxy.zap.extension.pscan.scanner.ScriptsPassiveScanner'
+java_import 'org.zaproxy.zap.extension.scripts.scanrules.ScriptsPassiveScanner'
 java_import 'org.parosproxy.paros.network.HttpMessage'
 java_import 'net.htmlparser.jericho.Source'
 java_import 'org.parosproxy.paros.view.View'
 
 class JRubyPassiveScript 
-  include Java::org.zaproxy.zap.extension.pscan.PassiveScript
+  include Java::org.zaproxy.zap.extension.scripts.scanrules.PassiveScript
 
   java_signature 'scan(ScriptsPassiveScanner, HttpMessage, Source)'
   def scan(ps, msg, src)