+++ addOns/scripts/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Added
+- The scan rule functionality of scripts was moved from the ZAP core to this add-on (Related to Issue 7105).
 
 ## [44] - 2023-12-19
 ### Changed